Add files via upload
First pulled tweets and analysis. Currently, the script only pulls tweets for a specific word or stock ticker which is in the code as 'TSLA' for testing. The eventual goal is to have a list of many stocks that are run through the analyzer on a semi-regular basis, returning an average score and finding increased scores.