completed but seperate DB connecting
this py file is siloed from the main file. The database connection is working correctly. returns the current sentiment score and the change from the last date of the database entry for each stock. still need to have this file become a function of the main file. Including using a larger set of tickers and returning data in the most efficient way.